# FE/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ChangeMind API",
    "description": "ASP.NET Core 10.0 API following Clean Architecture principles. Provides endpoints for fitness coaching, training programs, and personal health management.",
    "version": "1.0.0",
    "contact": {
      "name": "ChangeMind Team",
      "email": "ongguzel@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development (Gateway)"
    },
    {
      "url": "https://api.changemind.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/training-programs": {
      "post": {
        "tags": ["Training Programs"],
        "summary": "Create new training program (coach only)",
        "description": "Coaches create training programs for users. Requires JWT authentication.",
        "operationId": "CreateTrainingProgram",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingProgramRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training program created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}": {
      "get": {
        "tags": ["Training Programs"],
        "summary": "Get training program by ID",
        "operationId": "GetTrainingProgramById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Training program retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveProgramDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/active-program": {
      "get": {
        "tags": ["Training Programs"],
        "summary": "Get the user's active training program (Coach > Self priority)",
        "description": "Returns a single active program for the user. If a Coach-assigned active program exists, it is returned; otherwise the Self-created active program is returned (if any). Returns 404 when the user has neither.",
        "operationId": "GetUserActiveProgram",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Active program retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActiveProgramDetailDto" }
              }
            }
          },
          "404": {
            "description": "No active program for user",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
          }
        }
      }
    },
...
    "/api/training-programs/{id}/export": {
      "get": {
        "tags": ["Training Programs"],
        "summary": "Export training program as Excel (.xlsx)",
        "description": "Returns the program plan and full workout history as an .xlsx file. Each program day is rendered on its own sheet.",
        "operationId": "ExportTrainingProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Excel file stream",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}/activate": {
      "post": {
        "tags": ["Training Programs"],
        "summary": "Activate training program",
        "operationId": "ActivateTrainingProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Training program activated successfully"
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}/deactivate": {
      "post": {
        "tags": ["Training Programs"],
        "summary": "Deactivate training program",
        "operationId": "DeactivateTrainingProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Training program deactivated successfully"
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}/daily-program": {
      "put": {
        "tags": ["Training Programs"],
        "summary": "Update daily program exercises",
        "description": "Set weekly exercise schedule for training program",
        "operationId": "UpdateDailyProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDailyProgramRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Daily program updated successfully"
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}/progress": {
      "put": {
        "tags": ["Training Programs"],
        "summary": "Update training progress",
        "operationId": "UpdateProgress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Progress updated successfully"
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-programs/{id}/complete": {
      "post": {
        "tags": ["Training Programs"],
        "summary": "Mark training program as complete",
        "operationId": "CompleteTrainingProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Training program marked as complete"
          },
          "404": {
            "description": "Training program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTrainingProgramRequest": {
        "type": "object",
        "required": ["name", "userId", "durationWeeks", "difficulty"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "User ID to assign program to"
          },
          "durationWeeks": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "difficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          }
        }
      },
...
      "UpdateDailyProgramRequest": {
        "type": "object",
        "description": "Set weekly exercise schedule. Key format: Day-1, Day-2, etc.",
        "additionalProperties": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/ExerciseDetail"
          }
        },
        "example": {
          "Day-1": [
            {
              "exerciseId": "550e8400-e29b-41d4-a716-446655440000",
              "sets": 3,
              "reps": "10-12",
              "explanation": "Focus on form, not weight"
            }
          ]
        }
      },
      "ExerciseDetail": {
        "type": "object",
        "required": ["exerciseId", "sets", "reps"],
        "properties": {
          "exerciseId": {
            "type": "string",
            "format": "uuid"
          },
          "sets": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "reps": {
            "type": "string",
            "description": "Reps range (e.g., 10-12, 8-10, AMRAP)"
          },
          "explanation": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateProgressRequest": {
        "type": "object",
        "required": ["completedWeeks"],
        "properties": {
          "completedWeeks": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ActiveProgramDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "durationWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "createdByType": {
            "type": "string",
            "enum": ["Self", "Coach"],
            "description": "Indicates whether the user created this program themselves (`Self`) or a coach assigned it (`Coach`)."
          },
          "coachId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Coach ID if `createdByType=Coach`, otherwise null."
          },
          "coachName": {
            "type": "string",
            "nullable": true,
            "description": "Coach full name (FirstName + LastName). Null if `createdByType=Self`."
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "enum": ["Beginner", "Intermediate", "Advanced"]
          },
          "status": {
            "type": "string",
            "enum": ["InProgress", "Completed"]
          },
          "dailyExercises": {
            "type": "object",
            "description": "Weekly schedule (Day-1, Day-2, etc.) with exercises",
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProgramExerciseDetail"
              }
            }
          },
          "userId":     { "type": "string",  "format": "uuid" },
          "userAge":    { "type": "integer", "format": "int32", "nullable": true },
          "userHeight": { "type": "number",  "format": "decimal", "nullable": true, "description": "User height in cm" },
          "userWeight": { "type": "number",  "format": "decimal", "nullable": true, "description": "User weight in kg" },
          "userGender": { "type": "string",  "enum": ["Male", "Female", "Other"], "nullable": true }
        }
      },
...
      "ProgramExerciseDetail": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string",
            "format": "uuid"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "reps": {
            "type": "string"
          },
          "explanation": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://tools.ietf.org/html/rfc7807"
          },
          "title": {
            "type": "string",
            "example": "Not Found"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "User with ID 'xxx' not found."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/users/550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "One or more validation errors occurred."
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "errors": {
            "type": "object",
            "description": "Validation errors indexed by field name or index",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": ["Email is required", "Email format is invalid"],
              "password": ["Password must be at least 8 characters"]
            }
          }
        }
      }
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtained from /api/auth/login endpoint. Token validity: 15 minutes. Use Idempotency-Key header with POST /api/payments for idempotent payment processing."
      }
    }
  },
...
}